A patch from Sascha Rodekamp "The Job Sandbox should never be cached" (https://issues.apache.org/jira/browse/OFBIZ-3968) - OFBIZ-3968
Sascha: We noticed that OFBiz was sending cache clears over JMS every time the job sandbox is changed. That's not a problem with one OFBiz Instance but with 20, when every instance is sending JMS messages to each other ... you can imagine...

Scott: I'd agree with this, there is really no good reason to cache JobSandbox values even though I don't believe any code does this at the moment. On a side note I ran into some problems with the Job Poller when it's under a heavy load recently and as soon as I get a chance I'll be working to reproduce and fix the issues.


git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk/framework/service@1003418 13f79535-47bb-0310-9956-ffa450edef68