A patch from Varun Bhansaly "Issue receiving emails with attachment" https://issues.apache.org/jira/browse/OFBIZ-4413
Some Background : In our customization we have configured MCA to process incoming emails which is basically a custom service this service in turn calls service storeIncomingEmail, and rest of the custom code follows.
Now though we are able to receive emails (which can be seen in the communications in party manager), we have noticed that system does not some process the attachments and creates the link between communication event and content.

Analysis shows, MimeMessageWrapper.getAttachmentIndexes() while checking for content disposition, does a case sensitive comparison, which should rather be case insensitive.
This small change fixes the issue.



git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk/framework/service@1172213 13f79535-47bb-0310-9956-ffa450edef68