Bug fix: Applied patch from jira issue OFBIZ-4441. UtilXml firstChildElement should use getNodeName instead of getLocalName.
We have observed the same issue in past and provided a patch for the same. For more details please refer the commit r995384.

Fixing bug of reading node name. When we read first child element of the response of any third party integration for example ups etc, node name can't be read using "node.getLocalName()". So changing  node.getLocalName() --> node.getNodeName().

This time before I revert my changes I will wait for Bilgin's comment to understand the exact problem with SOAP call due to which he did this changes. Bilgin please post your comment on this as soon as you get a chance.
http://svn.apache.org/viewvc?view=revision&revision=894359 

Thanks Kiran, Sascha for the contribution. We were also observing this issue since quite long so for now it is better to put the code in trunk and releasebranch. 

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk/framework/service@1183314 13f79535-47bb-0310-9956-ffa450edef68