This partially reverts r1735569 to simplify and have less commented out parts.
Fixes "Comment out RMI related code because of the Java deserialization issue" - https://issues.apache.org/jira/browse/OFBIZ-6942

We decided to comment out as less as possible because once, in the start and both properties; the rmi part is off and the related test services are off there is no RMI related danger left (test services are not a danger but would fail during tests run). 
It's then easier for users who need RMI in their projects to have only to uncomment those and not digg everywhere.

[CVE-2016-2170] The infamous Java serialization vulnerability

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk/framework/service@1736083 13f79535-47bb-0310-9956-ffa450edef68