Implemented: Add the entity-auto CRUD services for RuntimeData (OFBIZ-9161)
At OFBIZ-9159 I added the entity-auto CRUD services for JobSandbox but 
w/o permission check. 

While working on this issue I decided it's better to have them and decided to 
also put them for JobSandbox CRUD services.

So here are the entity-auto CRUD services for RuntimeData and permission check 
for JobSandbox CRUD services


git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk/framework/service@1776558 13f79535-47bb-0310-9956-ffa450edef68